+++ NEWS.md
@@ -1,8 +1,15 @@
 # fusen 0.1.1
 
+**features**
+* Allow non-clean vignette name
+
+**documentation**
+* Add vignette to explain how to maintain a package with {fusen}
 * Add vignette to explain how to deal with data
-* Add tests for corner cases in Rmd templates
 * Added a `NEWS.md` file to track changes to the package.
+
+**tests**
+* Add tests for corner cases in Rmd templates
 
 # fusen 0.1.0